Refactored: Cleaned Styling, Added Price functionality to Sales, added Price to Products

+++ DB-functions/sql/productsTable.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS `products` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `product` varchar(30) NOT NULL,
   `category_id` int(11) DEFAULT NULL,
-  `price` int(11) DEFAULT NULL,
+  `price` decimal(5,2) DEFAULT NULL,
   PRIMARY KEY (`id`),
   UNIQUE KEY `product` (`product`),
   KEY `category_id` (`category_id`)
